pulled getThumbnail() logic out of index.js page
